requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==1.11.29
django-admin-interface==0.12.1
django-autoslug-iplweb==1.9.4.dev0
django-autotranslate==1.1.0
django-better-admin-arrayfield==1.0.5
django-bootstrap-breadcrumbs==0.9.2
django-cities==0.6
django-cities-light==3.5.0
django-ckeditor==5.9.0
django-colorfield==0.2.1
django-contrib-comments==1.9.2
django-extensions==2.2.9
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-model-utils==4.0.0
djangorestframework==3.11.0
filebrowser-safe==0.5.0
future==0.18.2
futures==3.1.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
goslate==1.5.1
grappelli-safe==0.5.2
httplib2==0.17.0
idna==2.9
Mezzanine==4.3.1
oauthlib==3.1.0
Pillow==7.0.0
polib==1.1.0
progressbar2==3.50.1
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-utils==2.4.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
swapper==1.1.2.post1
tqdm==4.43.0
tzlocal==2.0.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
